pyproject.toml
[project]
name = "static-site-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "decorator==5.2.1",
    "executing==2.2.0",
    "fastapi==0.115.12",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jsonpickle==4.1.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "networkx==3.4.2",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "prompt-toolkit==3.0.51",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pygal==3.0.5",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pytest>=8.4.1",
    "pyvis==0.3.2",
    "pyyaml>=6.0.2",
    "rdflib==7.1.4",
    "requests==2.32.3",
    "sniffio==1.3.1",
    "sparqlwrapper==2.0.0",
    "stack-data==0.6.3",
    "starlette==0.46.2",
    "tqdm>=4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "zipp==3.21.0",
]
